# swagger.yaml
openapi: 3.0.0
info:
  title: 4TP API
  description: 4TP API
  version: 0.0.1
servers:
  - url: "https://for-the-players.herokuapp.com/"
...
paths:
...
  /tickets:
    get:
      summary: Returns all tickets
      security:
        - BearerAuth: [ ]
      tags:
        - tickets
      operationId: listTickets
      responses:
        "400":
          description: "The input data was wrongly formatted"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchemaValidationError"
        "401":
          description: "Missing bearer token"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessageResponse"
        "403":
          description: "Forbidden"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessageResponse"
        "200":
          description: "The list of all tickets"
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    subject:
                      type: string
                      description: The subject of the ticket
                    _id:
                      type: string
                      description: The Mongo ID of the ticket
                    date:
                      type: string
                      format: date-time
                      description: The date of the ticket
                    category:
                      $ref: "#/components/schemas/TicketCategories"
                    matchId:
                      type: string
                      description: The Mongo ID of the linked match (optional)
                    status:
                      $ref: "#/components/schemas/TicketStatuses"
                    tournamentId:
                      type: string
                      description: The Mongo ID of the linked tournament (optional)
                    attachments:
                      type: array
                      items:
                        type: string
                        description: The url of the attachment
                    messages:
                      type: object
                      properties:
                        authorId:
                          type: string
                          description: The url of the attachment
                        message:
                          type: string
                          description: The message
                        date:
                          type: string
                          format: date-time
                          description: The date of the message
                        isAdmin:
                          type: boolean
                          description: If the message has been sent by an admin
                        _id:
                          type: string
                          description: The Id of the message
    post:
      summary: Create a new ticket
      security:
        - BearerAuth: [ ]
      tags:
        - tickets
      operationId: createTicket
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - subject
                - description
                - category
                - date
              properties:
                subject:
                  description: The ticket's subject
                  type: string
                description:
                  description: The ticket's description (this is going to be the first message)
                  type: string
                category:
                  $ref: "#/components/schemas/TicketCategories"
                matchId:
                  type: string
                  description: The linked match's ID
                tournamentId:
                  type: string
                  description: The linked tournament's ID
      responses:
        "400":
          description: "The input data was wrongly formatted"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchemaValidationError"
        "401":
          description: "Missing bearer token"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessageResponse"
        "403":
          description: "Forbidden"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessageResponse"
        "201":
          description: "Ticket added"
  /tickets/{ticketId}:
    parameters:
      - name: ticketId
        in: path
        required: true
        description: The ID of the ticket
        schema:
          $ref: "#/components/schemas/MongoID"
    get:
      summary: Returns a single ticket
      security:
        - BearerAuth: [ ]
      tags:
        - tickets
      operationId: getTicketDetail
      responses:
        "400":
          description: "The input data was wrongly formatted"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchemaValidationError"
        "401":
          description: "Missing bearer token"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessageResponse"
        "403":
          description: "Forbidden"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessageResponse"
        "200":
          description: "A single ticket"
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    subject:
                      type: string
                      description: The subject of the ticket
                    _id:
                      $ref: "#/components/schemas/MongoID"
                    createdAt:
                      type: string
                      format: date-time
                      description: When the ticket was created
                    messages:
                      type: array
                      items:
                        type: object
                        properties:
                          user:
                            type: object
                            properties:
                              username:
                                type: string
                              _id:
                                $ref: "#/components/schemas/MongoID"
                          message:
                            type: string
                          createdAt:
                            type: string
                            format: date-time
                          fromAdminPanel:
                            type: boolean
                            description: Indicates if this message has been sent from the admin panel or from the app
                    category:
                      $ref: "#/components/schemas/TicketCategories"
                    tournament:
                      type: object
                      description: Only present when the category is DISPUTE
                      properties:
                        _id:
                          $ref: "#/components/schemas/MongoID"
                        name:
                          type: string
                          description: The name of the linked tournament
                    match:
                      type: object
                      description: Only present when the category is DISPUTE
                      properties:
                        _id:
                          $ref: "#/components/schemas/MongoID"
                        acceptedDate:
                          type: string
                          format: date-time
                        teamOne:
                          type: object
                          properties:
                            name:
                              type: string
                            _id:
                              $ref: "#/components/schemas/MongoID"
                        teamTwo:
                          type: object
                          properties:
                            name:
                              type: string
                            _id:
                              $ref: "#/components/schemas/MongoID"
                    attachments:
                      type: array
                      items:
                        type: string
                        description: The url of the attachment
                    status:
                      $ref: "#/components/schemas/TicketStatuses"
    patch:
      summary: Update the ticket status
      security:
        - BearerAuth: [ ]
      tags:
        - tickets
      operationId: updateTicket
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  $ref: "#/components/schemas/TicketStatuses"
      responses:
        "400":
          description: "The input data was wrongly formatted"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchemaValidationError"
        "401":
          description: "Missing bearer token"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessageResponse"
        "403":
          description: "Forbidden"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessageResponse"
        200:
          description: "success"
  /tickets/{ticketId}/messages:
    parameters:
      - name: ticketId
        in: path
        required: true
        description: The ID of the ticket
        schema:
          $ref: "#/components/schemas/MongoID"
    post:
      summary: Add message to ticket
      security:
        - BearerAuth: [ ]
      tags:
        - tickets
      operationId: updateTicketMessage
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - message
                - fromAdminPanel
              properties:
                message:
                  type: string
                  description: The message body
                fromAdminPanel:
                  type: boolean
                  description: Indicates if the message is sent from the admin panel
      responses:
        "400":
          description: "The input data was wrongly formatted"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchemaValidationError"
        "401":
          description: "Missing bearer token"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessageResponse"
        "403":
          description: "Forbidden"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessageResponse"
        "201":
          description: "Ticket's message added"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
    SchemaValidationError:
      type: array
      items:
        type: object
        properties:
          value:
            type: string
            description: "The value that was received by the server"
          msg:
            type: string
            description: "Error message"
          param:
            type: string
            description: "Name of the field that caused the error"
          location:
            type: string
            description: "Where the field was that was received by the server"
    ErrorMessageResponse:
      type: object
      properties:
        errorMessage:
          type: string
...
    TicketStatuses:
      description: "The possible statuses for a ticket"
      type: string
      enum:
        - NEW
        - IN PROGRESS
        - SOLVED
        - DELETED
    TicketCategories:
      description: "The possible categories of a ticket"
      type: string
      enum:
        - DISPUTE
        - GENERAL_INQUIRY
        - PAYMENT
...
    MongoID:
      type: string
      maxLength: 24
      minLength: 24